last commit

+++ templates_c/5c040d59fdda6d937611a024094963d5545a024a_0.file.taskList.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 4.2.1, created on 2023-03-14 14:40:11
+/* Smarty version 4.2.1, created on 2023-03-14 15:38:58
   from 'E:\xampp\htdocs\web2\todolist-practice-1\templates\tasks\taskList.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.2.1',
-  'unifunc' => 'content_6410793bce2359_20148829',
+  'unifunc' => 'content_641087021ab920_17550239',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     '5c040d59fdda6d937611a024094963d5545a024a' => 
     array (
       0 => 'E:\\xampp\\htdocs\\web2\\todolist-practice-1\\templates\\tasks\\taskList.tpl',
-      1 => 1678801211,
+      1 => 1678804734,
       2 => 'file',
     ),
   ),
@@ -20,7 +20,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   array (
   ),
 ),false)) {
-function content_6410793bce2359_20148829 (Smarty_Internal_Template $_smarty_tpl) {
+function content_641087021ab920_17550239 (Smarty_Internal_Template $_smarty_tpl) {
 ?><div class="container text-center">
     <form action="home" method="get">
         <h1>Filter by type</h1>
@@ -54,12 +54,14 @@ if ($_prefixVariable1 == 0) {?>
             <div class="col">
                 <h2>Pending</h2>
 
+
             <?php
 $_from = $_smarty_tpl->smarty->ext->_foreach->init($_smarty_tpl, $_smarty_tpl->tpl_vars['tasks']->value, 'task');
 $_smarty_tpl->tpl_vars['task']->do_else = true;
 if ($_from !== null) foreach ($_from as $_smarty_tpl->tpl_vars['task']->value) {
 $_smarty_tpl->tpl_vars['task']->do_else = false;
 ?>
+
 
                 <?php if ($_smarty_tpl->tpl_vars['task']->value->status_id == 0) {?>
                         <div class="card m-2" style="width: 18rem;">
@@ -78,17 +80,21 @@ $_smarty_tpl->tpl_vars['task']->do_else = false;
                                 <p>Priority: <?php echo $_smarty_tpl->tpl_vars['task']->value->priority;?>
 </p>
 
-                    <?php if ($_smarty_tpl->tpl_vars['role']->value[0] != 0) {?>
-                                    <a href='deleteTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
-' class="btn btn-danger">Delete</a>
-                                    <a href='updateTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
-/1' class="btn btn-secondary">Take task</a>
 
+                    <?php if ($_smarty_tpl->tpl_vars['role']->value) {?>
+                        <?php if ($_smarty_tpl->tpl_vars['role']->value[0] != 0) {?>
+                                        <a href='deleteTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
+' class="btn btn-danger">Delete</a>
+                                        <a href='updateTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
+/1' class="btn btn-secondary">Take task</a>
+                        <?php }?>
                     <?php }?>
                             </div>
                         </div>
 
+
                 <?php }?>
+
 
             <?php
 }
@@ -97,12 +103,14 @@ $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
             <div class="col">
                 <h2>In course</h2>
 
+
             <?php
 $_from = $_smarty_tpl->smarty->ext->_foreach->init($_smarty_tpl, $_smarty_tpl->tpl_vars['tasks']->value, 'task');
 $_smarty_tpl->tpl_vars['task']->do_else = true;
 if ($_from !== null) foreach ($_from as $_smarty_tpl->tpl_vars['task']->value) {
 $_smarty_tpl->tpl_vars['task']->do_else = false;
 ?>
+
 
                 <?php if ($_smarty_tpl->tpl_vars['task']->value->status_id == 1) {?>
                         <div class="card m-2" style="width: 18rem;">
@@ -122,17 +130,21 @@ $_smarty_tpl->tpl_vars['task']->do_else = false;
                                 <p>Priority: <?php echo $_smarty_tpl->tpl_vars['task']->value->priority;?>
 </p>
 
+
                     <?php if ($_smarty_tpl->tpl_vars['task']->value->email == $_smarty_tpl->tpl_vars['username']->value) {?>
                                     <a href='updateTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
 /0' class="btn btn-secondary">Leave order</a>
                                     <a href='updateTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
 /2' class="btn btn-secondary">End task</a>
 
+
                     <?php }?>
+
 
                     <?php if ($_smarty_tpl->tpl_vars['role']->value[0] == 2) {?>
                                     <a href='deleteTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
 ' class="btn btn-danger">Delete</a>
+
 
 
                     <?php }?>
@@ -140,7 +152,9 @@ $_smarty_tpl->tpl_vars['task']->do_else = false;
                         </div>
 
 
+
                 <?php }?>
+
 
 
             <?php
@@ -150,12 +164,14 @@ $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
             <div class="col">
                 <h2>Finished</h2>
 
+
             <?php
 $_from = $_smarty_tpl->smarty->ext->_foreach->init($_smarty_tpl, $_smarty_tpl->tpl_vars['tasks']->value, 'task');
 $_smarty_tpl->tpl_vars['task']->do_else = true;
 if ($_from !== null) foreach ($_from as $_smarty_tpl->tpl_vars['task']->value) {
 $_smarty_tpl->tpl_vars['task']->do_else = false;
 ?>
+
 
                 <?php if ($_smarty_tpl->tpl_vars['task']->value->status_id == 2) {?>
                         <div class="card m-2" style="width: 18rem;">
@@ -173,6 +189,7 @@ $_smarty_tpl->tpl_vars['task']->do_else = false;
 </p>
                                 <p>Finished by <?php echo $_smarty_tpl->tpl_vars['task']->value->email;?>
 </p>
+
 
                     <?php if ($_smarty_tpl->tpl_vars['role']->value[0] != 0) {?>
                                     <a href='deleteTask/<?php echo $_smarty_tpl->tpl_vars['task']->value->id_task;?>
